@@ -3,7 +3,6 @@ package main
 import (
 	"crypto/tls"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"os"
 	"os/signal"
@@ -14,6 +13,7 @@ import (
 	"github.com/AdguardTeam/dnsproxy/proxy"
 	"github.com/AdguardTeam/dnsproxy/upstream"
 	"github.com/AdguardTeam/golibs/log"
+	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/ameshkov/dnscrypt/v2"
 	goFlags "github.com/jessevdk/go-flags"
 	"gopkg.in/yaml.v3"
@@ -159,7 +159,7 @@ type Options struct {
 	IPv6Disabled bool `yaml:"ipv6-disabled" long:"ipv6-disabled" description:"If specified, all AAAA requests will be replied with NoError RCode and empty answer" optional:"yes" optional-value:"true"`
 
 	// Transform responses that contain at least one of the given IP addresses into NXDOMAIN
-	BogusNXDomain []string `yaml:"bogus-nxdomain" long:"bogus-nxdomain" description:"Transform responses that contain at least one of the given IP addresses into NXDOMAIN. Can be specified multiple times."`
+	BogusNXDomain []string `yaml:"bogus-nxdomain" long:"bogus-nxdomain" description:"Transform the responses containing at least a single IP that matches specified addresses and CIDRs into NXDOMAIN.  Can be specified multiple times."`
 
 	// UDP buffer size value
 	UDPBufferSize int `yaml:"udp-buf-size" long:"udp-buf-size" description:"Set the size of the UDP buffer in bytes. A value <= 0 will use the system default."`
@@ -196,7 +196,7 @@ func main() {
 		if len(arg) > 13 {
 			if arg[:13] == "--config-path" {
 				fmt.Printf("Path: %s\n", arg[14:])
-				b, err := ioutil.ReadFile(arg[14:])
+				b, err := os.ReadFile(arg[14:])
 				if err != nil {
 					log.Fatalf("failed to read the config file %s: %v", arg[14:], err)
 				}
@@ -225,7 +225,7 @@ func run(options *Options) {
 		log.SetLevel(log.DEBUG)
 	}
 	if options.LogOutput != "" {
-		file, err := os.OpenFile(options.LogOutput, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0644)
+		file, err := os.OpenFile(options.LogOutput, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0o644)
 		if err != nil {
 			log.Fatalf("cannot create a log file: %s", err)
 		}
@@ -356,17 +356,19 @@ func initEDNS(config *proxy.Config, options *Options) {
 
 // initBogusNXDomain inits BogusNXDomain structure
 func initBogusNXDomain(config *proxy.Config, options *Options) {
-	if len(options.BogusNXDomain) > 0 {
-		bogusIP := []net.IP{}
-		for _, s := range options.BogusNXDomain {
-			ip := net.ParseIP(s)
-			if ip == nil {
-				log.Error("Invalid IP: %s", s)
-			} else {
-				bogusIP = append(bogusIP, ip)
-			}
+	if len(options.BogusNXDomain) == 0 {
+		return
+	}
+
+	for _, s := range options.BogusNXDomain {
+		subnet, err := netutil.ParseSubnet(s)
+		if err != nil {
+			log.Error("%s", err)
+
+			continue
 		}
-		config.BogusNXDomain = bogusIP
+
+		config.BogusNXDomain = append(config.BogusNXDomain, subnet)
 	}
 }
 
@@ -399,7 +401,7 @@ func initDNSCryptConfig(config *proxy.Config, options *Options) {
 		return
 	}
 
-	b, err := ioutil.ReadFile(options.DNSCryptConfigPath)
+	b, err := os.ReadFile(options.DNSCryptConfigPath)
 	if err != nil {
 		log.Fatalf("failed to read DNSCrypt config %s: %v", options.DNSCryptConfigPath, err)
 	}
@@ -569,11 +571,11 @@ func newTLSConfig(options *Options) (*tls.Config, error) {
 // form a certificate chain. On successful return, Certificate.Leaf will
 // be nil because the parsed form of the certificate is not retained.
 func loadX509KeyPair(certFile, keyFile string) (tls.Certificate, error) {
-	certPEMBlock, err := ioutil.ReadFile(certFile)
+	certPEMBlock, err := os.ReadFile(certFile)
 	if err != nil {
 		return tls.Certificate{}, err
 	}
-	keyPEMBlock, err := ioutil.ReadFile(keyFile)
+	keyPEMBlock, err := os.ReadFile(keyFile)
 	if err != nil {
 		return tls.Certificate{}, err
 	}
@@ -588,7 +590,7 @@ func loadServersList(sources []string) []string {
 	var servers []string
 
 	for _, source := range sources {
-		data, err := ioutil.ReadFile(source)
+		data, err := os.ReadFile(source)
 		if err != nil {
 			// Ignore errors, just consider it a server address
 			// and not a file
