@@ -2,15 +2,16 @@ package proxyutil
 
 import (
 	"encoding/binary"
-	"errors"
+	"fmt"
 	"io"
 	"net"
 
+	"github.com/AdguardTeam/golibs/errors"
 	"github.com/miekg/dns"
 )
 
-// ErrTooLarge - DNS message is larger than 64kb
-var ErrTooLarge = errors.New("DNS message is too large")
+// ErrTooLarge means that a DNS message is larger than 64KiB.
+const ErrTooLarge errors.Error = "dns message is too large"
 
 // DNSSize returns if buffer size *advertised* in the requests OPT record.
 // Or when the request was over TCP, we return the maximum allowed size of 64K.
@@ -32,13 +33,15 @@ func DNSSize(isUDP bool, r *dns.Msg) int {
 	return int(size)
 }
 
-// ReadPrefixed -- reads a DNS message with a 2-byte prefix containing message length
+// ReadPrefixed reads a DNS message with a 2-byte prefix containing message
+// length from conn.
 func ReadPrefixed(conn net.Conn) ([]byte, error) {
 	l := make([]byte, 2)
 	_, err := conn.Read(l)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("reading len: %w", err)
 	}
+
 	packetLen := binary.BigEndian.Uint16(l)
 	if packetLen > dns.MaxMsgSize {
 		return nil, ErrTooLarge
@@ -47,8 +50,9 @@ func ReadPrefixed(conn net.Conn) ([]byte, error) {
 	buf := make([]byte, packetLen)
 	_, err = io.ReadFull(conn, buf)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("reading msg: %w", err)
 	}
+
 	return buf, nil
 }
 
